% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_packages.R
\docType{package}
\name{IPM-package}
\alias{IPM}
\alias{IPM-package}
\title{IPM: Integrated Population Model}
\description{
Provides functions for simulating the dynamic population using two models, summarizing the output from MCMC simulations, and calaulating the credible interval by parametric bootstrap.
}
\author{
\strong{Maintainer}: Yiran Wang \email{y3577wan@uwaterloo.ca} (\href{https://orcid.org/0000-0003-1158-8367}{ORCID})

Other contributors:
\itemize{
  \item Yumeng Wang \email{y3496wan@uwaterloo.ca} [contributor]
}

}
\keyword{internal}
