% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_analyse.R
\name{ipm_analyse}
\alias{ipm_analyse}
\title{Analyse data for a simulation study}
\usage{
ipm_analyse(
  data,
  Plot = FALSE,
  model = "lcl",
  priors = NULL,
  maxtime = 5,
  unit = "mins",
  save = 20000L,
  chain = 3,
  moni,
  inits = NULL
)
}
\arguments{
\item{data}{an nlists object, simulation dataset returned by ipm_sim.}

\item{Plot}{a flag, indicates whether to save the traceplot and the density plot for MCMC outputs.}

\item{model}{a string, specifying the model of interest. Default is "lcl", which means using composite likelihood to analyse the simulation from true joint likelihood. Similarly, you can use "ll" or "clcl".}

\item{priors}{a string of code to set the prior.}

\item{maxtime}{a scalar, specifying the maximum time to spend on analysis.}

\item{unit}{a character string specifying the units of time for \code{max.time}. See \code{difftime}.}

\item{save}{a scalar, the number of (potentially thinned) samples to save.}

\item{chain}{a scalar, the number of MCMC chains.}

\item{moni}{a vector of string or a string, specifying the parameters want to monitor.}

\item{inits}{a list, specifying the initial values of parameters.}
}
\value{
a mcmcrs object, the MCMC outputs, and plots (If plot=TRUE).
}
\description{
Using the model of interest to analyse the simulation by the model of interest.
}
