% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_analyse_cl.R
\name{ipm_analyse_cl}
\alias{ipm_analyse_cl}
\title{Analyse data using composite likelihood for a simulation study}
\usage{
ipm_analyse_cl(
  data,
  Plot = FALSE,
  priors = NULL,
  maxtime = 5,
  unit = "mins",
  save = 20000L,
  chain = 3
)
}
\arguments{
\item{data}{an nlists object, simulation dataset returned by ipm_sim_cl.}

\item{Plot}{a flag, indicates whether to save the traceplot and the density plot for MCMC outputs.}

\item{priors}{a string of code to set the prior.}

\item{maxtime}{a scalar, specifying the maximum time to spend on analysis.}

\item{unit}{a character string specifying the units of time for \code{max.time}. See \code{difftime}.}

\item{save}{a scalar, the number of (potentially thinned) samples to save.}

\item{chain}{a scalar, the number of MCMC chains.}
}
\value{
a mcmcrs object, the MCMC outputs, and plots (If plot=TRUE).
}
\description{
Using the composite likelihood model to analyse the simulation by composite likelihood model.
}
