% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_bootstrap.R
\name{ipm_bootstrap}
\alias{ipm_bootstrap}
\title{Parametric Bootstrap Estimation}
\usage{
ipm_bootstrap(
  result,
  data = NULL,
  type = "l",
  real = NULL,
  priors = NULL,
  times = 1000,
  maxtime = 5,
  unit = "mins",
  save = 20000L,
  chain = 3
)
}
\arguments{
\item{result}{an mcmcrs object, the result from ipm_analyse.}

\item{data}{an nlists object, simulation dataset returned by ipm_sim.}

\item{type}{a string, specifying the model used to analyse, commonly "l", "cl" and "clcl".}

\item{real}{an list object, specifying the real value using for simulation.}

\item{priors}{a string of code to set the prior.}

\item{times}{a scalar, specifying the number of samples.}

\item{maxtime}{a scalar, specifying the maximum time to spend on analysis.}

\item{unit}{a character string specifying the units of time for \code{max.time}. See \code{difftime}.}

\item{save}{a scalar, the number of (potentially thinned) samples to save.}

\item{chain}{a scalar, the number of MCMC chains.}
}
\value{
a list object, containing credible interval, RMSE and coverage probability.
}
\description{
Using parametric bootstrap to compute credible interval.
}
