% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_coverage.R
\name{ipm_coverage}
\alias{ipm_coverage}
\title{Coverage probability}
\usage{
ipm_coverage(result, real = NULL, times = 1000)
}
\arguments{
\item{result}{an mcmcrs object, the result from ipm_analyse or ipm_analyse_l.}

\item{real}{an list object, specifying the real value using for simulation.}

\item{times}{a scalar, specifying the number of samples.}
}
\value{
a mcmcrs object, the MCMC outputs.
}
\description{
Calculate the coverage probability.
}
