% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_rmse.R
\name{ipm_rmse}
\alias{ipm_rmse}
\title{Root mean square error}
\usage{
ipm_rmse(result, real = NULL, times = 1000)
}
\arguments{
\item{result}{an mcmcrs object, the result from ipm_analyse or ipm_analyse_l.}

\item{real}{an list object, specifying the real value using for simulation.}

\item{times}{a scalar, specifying the number of samples.}
}
\value{
a vector, the RMSE of each parameter.
}
\description{
Calculate the root mean square error.
}
