% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_sim_cl.R
\name{ipm_sim_cl}
\alias{ipm_sim_cl}
\title{Composite likelihood simulation}
\usage{
ipm_sim_cl(
  K = 4,
  N.1 = 500,
  sigma = 30,
  p = rep(0.8, 4),
  f = rep(2, 4),
  phi = rep(0.9, 3),
  R = NULL
)
}
\arguments{
\item{K}{a scalar, number of years.}

\item{N.1}{a scalar,size of initial population.}

\item{sigma}{a scalar, standard deviation of count data.}

\item{p}{K x 1 vector, recapture probability of every year.}

\item{f}{K x 1 vector, fecundity rate of every year.}

\item{phi}{(K-1) x 1 vector, mortality rate of every year.}

\item{R}{(K-1) x 1 vector, number of release in every year.}
}
\value{
data, the simulation result.
}
\description{
Simulate population by the composite likelihood model
}
