% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_summary.R
\name{ipm_summary}
\alias{ipm_summary}
\title{Summary of the result from analysis}
\usage{
ipm_summary(result)
}
\arguments{
\item{result}{an mcmcrs object, the result from ipm_analyse or ipm_analyse_l.}
}
\value{
a matirx, with summary staitstics.
}
\description{
Summary of the result from analysis
}
