% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_coverage.R
\name{ipm_coverage}
\alias{ipm_coverage}
\title{Coverage probability}
\usage{
ipm_coverage(result, real = NULL, times = 1000, type = "l")
}
\arguments{
\item{result}{an mcmcrs object, the result from ipm_analyse or ipm_analyse_l or ipm_analyse_cl.}

\item{real}{an list object, specifying the real value using for simulation.}

\item{times}{a scalar, specifying the number of samples.}

\item{type}{a string, specifying if it's from ipm_analyse_cl.}
}
\value{
a vector, the coverage probability of each parameter.
}
\description{
Calculate the coverage probability.
}
